+++ changelog.md
@@ -11,13 +11,14 @@
 * `contains_column`, `is_empty_by_name` and `is_empty` functions for `Row`.
 * `Display` implementation for public enums.
 * Missing traits for `Consistency` and `ColType`.
+* New `PreparedMetadataFlags`.
 
 ### Changed
 
 * Rewritten transport layer for massive performance improvements (including removing `bb8`). This
   involves changing a large portion of public API related to transport and server events.
 * Rewritten event mechanism - now you can subscribe to server events via `create_event_receiver()` in `Session`.
-* Replaced `QueryFlags` and `frame::Flags` vectors with bitflags.
+* Replaced `RowsMetadataFlag`, `QueryFlags` and `frame::Flags` vectors with bitflags.
 * Changed `Target` and `ChangeType` enums to `SchemaChangeTarget` and `SchemaChangeType`.
 * The `varint` type now uses `num::BigInt` representation (this implies `Decimal` also uses "big" types).
 * Removed `unstable-dynamic-cluster` feature, since it wasn't working as expected and introduced performance penalty.